♻️[Refactor] deprecated foregrounColor 수정
