+++ ReMU/ReMU/Common/UIComponents/Button/PrimaryButton.swift
@@ -42,7 +42,7 @@ struct PrimaryButton: View {
         } label: {
             Text(title)
                 .font(.pt16)
-                .foregroundColor(textColor)
+                .foregroundStyle(textColor)
                 .frame(maxWidth: .infinity) // View에서 패딩을 이용해서 가로 길이 설정!
                 .padding(.vertical, 16)
                 .background(isDisabled ? .purpleC495E0.opacity(0.4) : backgroundColor)
