+++ ReMU/ReMU/Models/DTO/Response/Result/CheckResultResponse.swift
@@ -12,4 +12,12 @@ struct CheckResultResponse: Decodable {
     let travelEmojiImageName: String
     let overallContent: String
     let aiFeedback: String?
+    let reviews: [CheckReviewResponse]
+}
+
+struct CheckReviewResponse: Decodable {
+    let reviewId: Int
+    let title: String
+    let content: String
+    let isResolutionFulfilled: Bool
 }
